requirements.txt
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Deprecated==1.2.9
Flask==1.1.2
Flask-Testing==0.8.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
multi-key-dict==2.0.3
openbrokerapi==4.0.2
packaging==20.3
pbr==5.4.5
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
PyGithub==1.47
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.13.0
python-jenkins==1.7.0
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.1.0
